Split simplify.cc into more files to improve build time
